requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
Django==5.2
flatbuffers==25.1.24
gast==0.6.0
google-pasta==0.2.0
grpcio==1.70.0
h11==0.14.0
idna==3.10
jiter==0.9.0
kagglehub==0.3.6
keras==3.8.0
libclang==18.1.1
lxml==5.3.0
lz4==4.4.4
ml-dtypes==0.4.1
mtcnn==1.0.0
namex==0.0.8
numpy==2.2.3
openai==1.72.0
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.0
outcome==1.3.0.post0
pillow==11.0.0
pycparser==2.22
PySocks==1.7.1
python-docx==1.1.2
python-pptx==1.0.2
requests==2.32.3
selenium==4.29.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
trio==0.29.0
trio-websocket==0.12.1
typing_extensions==4.12.2
tzdata==2025.2
undetected-chromedriver==3.5.5
urllib3==2.3.0
websocket-client==1.8.0
websockets==15.0
wsproto==1.2.0
XlsxWriter==3.2.0
